+++ docs/guides/admin/docs/releasenotes.md
@@ -44,7 +44,8 @@ Configuration changes
   controlled by Opencast and the routing through capture agents which existed for historical reasons was just an
   additional source for errors. If you rely on the old behavior, it can be configured in
   `etc/org.opencastproject.ingest.impl.IngestServiceImpl.cfg`.
-- The Paella player now respects all tracks published to engage instead of tracks with the flavor `*/delivery` only.
+- By default, the Paella player now respects all tracks published to engage instead of having a hard-coded filter for
+  tracks with the sub-flavor `delivery` only.
 - The structure of the configuration files concerning URL signing has changed.
   See [here](./configuration/stream-security.md). The affected files are:
     - `etc/org.opencastproject.security.urlsigning.filter.UrlSigningFilter.cfg `